# Clean up and optimize Frodo implementation 