--- a/src/kex_lwe_frodo/lwe_noise.c
+++ b/src/kex_lwe_frodo/lwe_noise.c
@@ -5,6 +5,7 @@
 #include <string.h>
 
 #include <oqs/rand.h>
+#include <oqs/common.h>
 
 #include "local.h"
 
@@ -14,9 +15,8 @@ static void lwe_sample_n_inverse_8(uint16_t *s, const size_t n, const uint8_t *c
 	 * timing: the CDF table is ingested for every sample.
 	 */
 
-	size_t rndlen = n;
-	uint8_t rndvec[rndlen];
-	OQS_RAND_n(rand, rndvec, rndlen);
+	uint8_t rndvec[n];
+	OQS_RAND_n(rand, rndvec, sizeof(rndvec));
 
 	for (size_t i = 0; i < n; ++i) {
 		uint8_t sample = 0;
@@ -32,8 +32,7 @@ static void lwe_sample_n_inverse_8(uint16_t *s, const size_t n, const uint8_t *c
 		// Assuming that sign is either 0 or 1, flips sample iff sign = 1
 		s[i] = ((-sign) ^ sample) + sign;
 	}
-
-	memset(rndvec, 0, rndlen);
+	OQS_MEM_cleanse(rndvec, sizeof(rndvec));
 }
 
 static void lwe_sample_n_inverse_12(uint16_t *s, const size_t n, const uint16_t *cdf_table, const size_t cdf_table_len, OQS_RAND *rand) {
@@ -42,10 +41,8 @@ static void lwe_sample_n_inverse_12(uint16_t *s, const size_t n, const uint16_t 
 	 * timing: the CDF table is ingested for every sample.
 	 */
 
-	size_t rndlen = 3 * ((n + 1) / 2);  // 12 bits of unif randomness per output element
-
-	uint8_t rnd[rndlen];
-	OQS_RAND_n(rand, rnd, rndlen);
+	uint8_t rnd[3 * ((n + 1) / 2)]; // 12 bits of unif randomness per output element
+	OQS_RAND_n(rand, rnd, sizeof(rnd));
 
 	for (size_t i = 0; i < n; i += 2) {  // two output elements at a time
 		uint8_t *pRnd = (rnd + 3 * i / 2);
@@ -74,8 +71,7 @@ static void lwe_sample_n_inverse_12(uint16_t *s, const size_t n, const uint16_t 
 			s[i + 1] = ((-sign2) ^ sample2) + sign2;
 		}
 	}
-
-	memset(rnd, 0, rndlen);
+	OQS_MEM_cleanse(rnd, sizeof(rnd));
 }
 
 static void lwe_sample_n_inverse_16(uint16_t *s, const size_t n, const uint16_t *cdf_table, const size_t cdf_table_len, OQS_RAND *rand) {
@@ -84,9 +80,8 @@ static void lwe_sample_n_inverse_16(uint16_t *s, const size_t n, const uint16_t 
 	 * timing: the CDF table is ingested for every sample.
 	 */
 
-	size_t rndlen = 2 * n;
-	uint16_t rndvec[rndlen];
-	OQS_RAND_n(rand, (uint8_t *) rndvec, rndlen);
+	uint16_t rndvec[ n ];
+	OQS_RAND_n(rand, (uint8_t *) rndvec, sizeof(rndvec));
 
 	for (size_t i = 0; i < n; ++i) {
 		uint8_t sample = 0;
@@ -102,8 +97,7 @@ static void lwe_sample_n_inverse_16(uint16_t *s, const size_t n, const uint16_t 
 		// Assuming that sign is either 0 or 1, flips sample iff sign = 1
 		s[i] = ((-sign) ^ sample) + sign;
 	}
-
-	memset(rndvec, 0, rndlen);
+	OQS_MEM_cleanse(rndvec, sizeof(rndvec));
 }
 
 void oqs_kex_lwe_frodo_sample_n(uint16_t *s, const size_t n, struct oqs_kex_lwe_frodo_params *params, OQS_RAND *rand) {