… src/kex_lwe_frodo/lwe_macrify.c
… src/kex_lwe_frodo/recommended.h