This is a WIP so don't commit it yet.

I noticed that `#define` (from #45 ) is used for parameters in some places and in other places the params stored in the `OQS_KEX` struct are used.  It would probably be a good idea to declare all params in one place for consistency.  This would have the additional benefit of having the values known at runtime throughout the implementation.

Currently the speed up on my machine is as follows:

Before:
```
Operation                           Iterations  Total time (s)  Time(us): mean      pop. stdev    Cycles: mean      pop. stdev
LWE Frodo recommended
alice 0                                    963           1.001        1039.050          71.687         2592656          178724
bob                                        694           1.000        1441.037          78.622         3596023          195931
alice 1                                   8574           1.000         116.633          12.134          290468           30179
```

After:

```
Operation                           Iterations  Total time (s)  Time(us): mean      pop. stdev    Cycles: mean      pop. stdev
LWE Frodo recommended
alice 0                                   1140           1.001         877.693          65.233         2189929          162453
bob                                        789           1.001        1268.984          51.522         3166522          128238
alice 1                                   8563           1.000         116.789          13.649          290883           33964
```