@aparent 2016-11-25
Unpack looks like it currently takes up around 15% of the runtime.  It could be made much faster if it was less generic.  Not sure that is worth doing right now though.