@aparent 2016-12-05
Will merge this soon unless anyone sees any issues with it?